factoring out rerun routines to fba_rerun_io.py
